chore: rename folders to follow restful patterns
