+++ web/src/app/(backend)/api/materias/[id]/route.ts
