+++ web/src/app/(backend)/api/materias/route.ts
