+++ web/src/app/(backend)/api/materias/slug/[slug]/route.ts
