+++ web/tests/integration/api/materias/index.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach, Mock } from 'vitest'
 import * as materiaService from '@/backend/services/materia'
-import { GET, POST } from '@/backend/api/materia/route'
+import { GET, POST } from '@/backend/api/materias/route'
 import { getMateriasMock, postMateriaMock } from '../../mocks/materia'
 import { setCurrentRole } from '../../mocks/auth'
 import { createRequest } from '../../mocks/requests'
@@ -10,7 +10,7 @@ vi.mock('@/backend/services/materia', () => ({
   createMateria: vi.fn(),
 }))
 
-describe('GET /api/materia', () => {
+describe('GET /api/materias', () => {
   beforeEach(() => {
     vi.clearAllMocks();
     setCurrentRole(null);
@@ -27,7 +27,7 @@ describe('GET /api/materia', () => {
   });
 });
 
-describe('POST /api/materia', () => {
+describe('POST /api/materias', () => {
   beforeEach(() => {
     vi.clearAllMocks();
     setCurrentRole(null); // Reset auth context
