+++ web/tests/integration/api/users/index.test.ts
@@ -8,9 +8,9 @@ vi.mock('@/backend/services/user', () => ({
   findUserByEmail: vi.fn(),
 }))
 
-import { POST } from '@/backend/api/user/route'
+import { POST } from '@/backend/api/users/route'
 
-describe('POST /api/user', () => {
+describe('POST /api/users', () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
